Moved to a separate test
