+++ uitest/src/test/java/com/vaadin/tests/components/tabsheet/TabsheetNotEnoughHorizontalSpaceTest.java
@@ -2,9 +2,6 @@ package com.vaadin.tests.components.tabsheet;
 
 import java.io.IOException;
 
-import com.vaadin.testbench.elements.ButtonElement;
-import com.vaadin.testbench.elements.TabSheetElement;
-import com.vaadin.tests.tb3.SingleBrowserTest;
 import org.junit.Test;
 
 import com.vaadin.testbench.By;
@@ -18,7 +15,7 @@ import com.vaadin.tests.tb3.MultiBrowserTest;
  *
  * @author Vaadin Ltd
  */
-public class TabsheetNotEnoughHorizontalSpaceTest extends SingleBrowserTest {
+public class TabsheetNotEnoughHorizontalSpaceTest extends MultiBrowserTest {
 
     @Test
     public void testThatTabScrollButtonsAreRenderedCorrectly()
@@ -31,13 +28,4 @@ public class TabsheetNotEnoughHorizontalSpaceTest extends SingleBrowserTest {
         compareScreen("init");
     }
 
-    @Test
-    public void deleteSelectedTab() {
-        openTestURL();
-        $(ButtonElement.class).caption("Select last tab").first().click();
-        assertElementPresent(By.className("v-tabsheet-tabitem"));
-        $(TabSheetElement.class).first().waitForVaadin();
-        $(ButtonElement.class).caption("Remove all tabs").first().click();
-        assertElementNotPresent(By.className("v-tabsheet-tabitem"));
-    }
 }
